+++ apps/hyperlocalise-web/src/workflows/email-translation.ts
@@ -1,6 +1,3 @@
-import { isUtf8 } from "node:buffer";
-import { createHash } from "node:crypto";
-
 import { Sandbox } from "@vercel/sandbox";
 import { and, eq } from "drizzle-orm";
 import { Resend } from "resend";
@@ -8,17 +5,11 @@ import { getWorkflowMetadata } from "workflow";
 
 import { db, schema } from "@/lib/database";
 import { env } from "@/lib/env";
-import { createLogger } from "@/lib/log";
-import {
-  type AttachmentContent,
-  inferAttachmentContentType,
-  toAttachmentBuffer,
-  toBase64AttachmentContent,
-} from "@/lib/resend/attachments";
+import { inferAttachmentContentType, toBase64AttachmentContent } from "@/lib/resend/attachments";
+import { getTranslatedFileDiagnostics } from "@/lib/translation/diagnostics";
 import type { EmailAgentTask, EmailAgentTaskAttachment } from "@/lib/workflow/types";
 
 const sandboxTimeoutMs = 10 * 60 * 1000;
-const logger = createLogger("email-translation-workflow");
 
 async function createTranslationSandbox(): Promise<{ sandboxId: string }> {
   "use step";
@@ -171,52 +162,6 @@ async function readTranslatedFile(sandboxId: string, outputFile: string): Promis
   return Buffer.from(content);
 }
 
-type TranslatedFileDiagnostics = {
-  filename: string;
-  byteLength: number;
-  sha256: string;
-  firstBytesHex: string;
-  contentType: string;
-  isUtf8: boolean;
-  jsonParseOk: boolean | null;
-  jsonParseError: string | null;
-};
-
-export async function getTranslatedFileDiagnostics(
-  content: AttachmentContent,
-  filename: string,
-): Promise<TranslatedFileDiagnostics> {
-  "use step";
-
-  const fileContent = toAttachmentBuffer(content);
-  const dotIndex = filename.lastIndexOf(".");
-  const ext = dotIndex === -1 ? "" : filename.slice(dotIndex).toLowerCase();
-  const isJsonLike = ext === ".json" || ext === ".jsonc";
-  let jsonParseOk: boolean | null = null;
-  let jsonParseError: string | null = null;
-
-  if (isJsonLike) {
-    try {
-      JSON.parse(fileContent.toString("utf8"));
-      jsonParseOk = true;
-    } catch (error) {
-      jsonParseOk = false;
-      jsonParseError = error instanceof Error ? error.message : String(error);
-    }
-  }
-
-  return {
-    filename,
-    byteLength: fileContent.byteLength,
-    sha256: createHash("sha256").update(fileContent).digest("hex"),
-    firstBytesHex: fileContent.subarray(0, 16).toString("hex"),
-    contentType: inferAttachmentContentType(filename),
-    isUtf8: isUtf8(fileContent),
-    jsonParseOk,
-    jsonParseError,
-  };
-}
-
 async function logTranslatedFileDiagnostics(
   task: EmailAgentTask,
   attachment: EmailAgentTaskAttachment,
@@ -227,16 +172,13 @@ async function logTranslatedFileDiagnostics(
 
   const diagnostics = await getTranslatedFileDiagnostics(translatedContent, outputFilename);
 
-  logger.info(
-    {
-      requestId: task.requestId,
-      attachmentId: attachment.id,
-      sourceFilename: attachment.filename,
-      targetLocale: task.parameters.translate.targetLocale,
-      diagnostics,
-    },
-    "translated email attachment diagnostics",
-  );
+  console.info("[email-translation-workflow] translated email attachment diagnostics", {
+    requestId: task.requestId,
+    attachmentId: attachment.id,
+    sourceFilename: attachment.filename,
+    targetLocale: task.parameters.translate.targetLocale,
+    diagnostics,
+  });
 }
 
 function shellQuote(value: string): string {
